move socket to compDidMount for App
